+++ src/styled-components/AuthStyle.ts
@@ -60,3 +60,9 @@ export const Switcher = styled.span`
     text-decoration-line: none;
   }
 `;
+
+export const ErrorMessage = styled.div`
+  color : red;
+  font-size : 16px;
+ text-align : center;
+`
